Simplify theming for `Application`
