+++ winit/src/multi_window.rs
@@ -20,8 +20,9 @@ use crate::runtime::command::{self, Command};
 use crate::runtime::multi_window::Program;
 use crate::runtime::user_interface::{self, UserInterface};
 use crate::runtime::Debug;
-use crate::style::application::StyleSheet;
 use crate::{Clipboard, Error, Proxy, Settings};
+
+pub use crate::application::{default, Appearance, Style};
 
 use std::collections::HashMap;
 use std::mem::ManuallyDrop;
@@ -41,7 +42,7 @@ use std::time::Instant;
 /// can be toggled by pressing `F12`.
 pub trait Application: Program
 where
-    Self::Theme: StyleSheet,
+    Style<Self::Theme>: Default,
 {
     /// The data needed to initialize your [`Application`].
     type Flags;
@@ -66,8 +67,8 @@ where
     fn theme(&self, window: window::Id) -> Self::Theme;
 
     /// Returns the `Style` variation of the `Theme`.
-    fn style(&self) -> <Self::Theme as StyleSheet>::Style {
-        Default::default()
+    fn style(&self, theme: &Self::Theme) -> Appearance {
+        Style::default().resolve(theme)
     }
 
     /// Returns the event `Subscription` for the current state of the
@@ -108,7 +109,7 @@ where
     A: Application + 'static,
     E: Executor + 'static,
     C: Compositor<Renderer = A::Renderer> + 'static,
-    A::Theme: StyleSheet,
+    Style<A::Theme>: Default,
 {
     use winit::event_loop::EventLoopBuilder;
 
@@ -349,7 +350,7 @@ async fn run_instance<A, E, C>(
     A: Application + 'static,
     E: Executor + 'static,
     C: Compositor<Renderer = A::Renderer> + 'static,
-    A::Theme: StyleSheet,
+    Style<A::Theme>: Default,
 {
     use winit::event;
     use winit::event_loop::ControlFlow;
@@ -819,7 +820,7 @@ fn build_user_interface<'a, A: Application>(
     id: window::Id,
 ) -> UserInterface<'a, A::Message, A::Theme, A::Renderer>
 where
-    A::Theme: StyleSheet,
+    Style<A::Theme>: Default,
 {
     debug.view_started();
     let view = application.view(id);
@@ -847,7 +848,7 @@ fn update<A: Application, C, E: Executor>(
     ui_caches: &mut HashMap<window::Id, user_interface::Cache>,
 ) where
     C: Compositor<Renderer = A::Renderer> + 'static,
-    A::Theme: StyleSheet,
+    Style<A::Theme>: Default,
 {
     for message in messages.drain(..) {
         debug.log_message(&message);
@@ -890,7 +891,7 @@ fn run_command<A, C, E>(
     A: Application,
     E: Executor,
     C: Compositor<Renderer = A::Renderer> + 'static,
-    A::Theme: StyleSheet,
+    Style<A::Theme>: Default,
 {
     use crate::runtime::clipboard;
     use crate::runtime::system;
@@ -1216,8 +1217,8 @@ pub fn build_user_interfaces<'a, A: Application, C: Compositor>(
     mut cached_user_interfaces: HashMap<window::Id, user_interface::Cache>,
 ) -> HashMap<window::Id, UserInterface<'a, A::Message, A::Theme, A::Renderer>>
 where
-    A::Theme: StyleSheet,
     C: Compositor<Renderer = A::Renderer>,
+    Style<A::Theme>: Default,
 {
     cached_user_interfaces
         .drain()
